nexus-sdk/foxmq/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nexus-sdk"
version = "0.1.0"
description = "Nexus Agentic Commerce SDK (FoxMQ via Python)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "paho-mqtt>=2.0.0",
    "pydantic>=2.0.0"
]

[tool.setuptools.packages.find]
where = ["src"]
